backend/requirements.txt
fastapi
uvicorn
python-multipart
httpx
sqlalchemy
python-dotenv
psycopg2-binary
